% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Account}
\alias{Account}
\title{Account}
\usage{
Account(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-account/ for more information.
}
