% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Alert}
\alias{Alert}
\title{Alert}
\usage{
Alert(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Alert/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("Alert Example"),

  # Alert -------------------------------------------------------------------
  tags$text("Alert"),
  p(),
  Alert(severity = "info",
        AlertTitle(tags$h5("Aviso")),
        tags$h6("This is an Alert")),
  tags$br()

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
