% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AppBar}
\alias{AppBar}
\title{AppBar}
\usage{
AppBar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/App-Bar/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("AppBar Example"),

  # AppBar ------------------------------------------------------------------
  AppBar(
    position="fixed", sx = 1.5,
    Toolbar(enableColorOnDark="TRUE",
            IconButton(size="large", edge="start", color="inherit", sx=2, BarChartIcon()),
            Typography(variant="h6", "News"),
            Button("LOGIN", color= "inherit")
    )
  ),

  tags$div(style="height:65px;"),
  hr(),

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
