% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{AppProvider}
\alias{AppProvider}
\title{AppProvider}
\usage{
AppProvider(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-app-provider/ for more information.
}
