% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Card}
\alias{Card}
\title{Card}
\usage{
Card(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Card/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("Card Example"),

  # Card --------------------------------------------------------------------
  tags$text("Card"),
  Card(
    CardContent("Hello world!")
  ),
  tags$br(),

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
