% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Checkbox}
\alias{Checkbox}
\title{Checkbox}
\usage{
Checkbox(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Checkbox/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("Checkbox Example"),

  # Checkbox Input ----------------------------------------------------------
  tags$text("Checkbox Input"),
  p(),
  Checkbox(defaultChecked =T, name= "Peach", size="large", color="secondary"),
  tags$br(),

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
