% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.R
\name{custom4mui_example}
\alias{custom4mui_example}
\title{Component example}
\usage{
custom4mui_example(...)
}
\arguments{
\item{...}{ui taglist to render}
}
\description{
Used for documentation purposes to show shiny examples.
}
