% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{darkScrollbar}
\alias{darkScrollbar}
\title{darkScrollbar (NEW)}
\usage{
darkScrollbar(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Dialog/ for more information.
}
