% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DashboardLayout}
\alias{DashboardLayout}
\title{DashboardLayout}
\usage{
DashboardLayout(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-dashboard-layout/ for more information.
}
