% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DataGrid}
\alias{DataGrid}
\title{#' createTheme (NEW)
#' @export
#' @md
#' @param ... args to pass to element
#' @description
#' Visit the documentation at https://mui.com/api/Theme-Provider/ for more information.
createTheme <- component("createTheme")
#' useTheme (NEW)
#' @export
#' @md
#' @param ... args to pass to element
#' @description
#' Visit the documentation at https://mui.com/api/Theme-Provider/ for more information.
useTheme <- component("useTheme")
DataGrid}
\usage{
DataGrid(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
#' createTheme (NEW)
#' @export
#' @md
#' @param ... args to pass to element
#' @description
#' Visit the documentation at https://mui.com/api/Theme-Provider/ for more information.
createTheme <- component("createTheme")
#' useTheme (NEW)
#' @export
#' @md
#' @param ... args to pass to element
#' @description
#' Visit the documentation at https://mui.com/api/Theme-Provider/ for more information.
useTheme <- component("useTheme")
DataGrid
}
