% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Fab}
\alias{Fab}
\title{Fab}
\usage{
Fab(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Fab/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("Floating Action Button Example"),

  # Floating Action Button --------------------------------------------------
  tags$text("Floating Action Button"),
  p(),
  Fab(color="primary", size="medium", BarChartIcon()),
  tags$br(),

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
