% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Icon}
\alias{Icon}
\title{Icon (BLOCKED)}
\usage{
Icon(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Icon/ for more information.
}
