% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ListItem}
\alias{ListItem}
\title{ListItem}
\usage{
ListItem(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/List-Item/ for more information.
}
