% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Paper}
\alias{Paper}
\title{Paper}
\usage{
Paper(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Paper/ for more information.
}
\examples{
library(shiny)
library(custom4mui)


# Define UI for application that draws a histogram
ui <- fluidPage(

  # Paper example ------------------------------------------------------------------
  tags$text("Paper"),
  p(),
  Paper(elevation = 3,
        tags$br(),
        "Esto es una prueba",
        tags$br(),
  ),
  tags$br(),

)


# Define server logic
server <- function(input, output) {

}

# Run the application
shinyApp(ui = ui, server = server)
}
