% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{persistence}
\alias{persistence}
\title{persistence}
\usage{
persistence(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-persistent-state/ for more information.
}
