% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Rating}
\alias{Rating}
\title{Rating}
\usage{
Rating(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Rating/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("Rating Example"),

  # Rating Input ------------------------------------------------------------
  tags$text("Rating Input"),
  p(),
  Rating("rating",
         name="simple-controlled",
         value=3.5,
         precision=0.5,
         size="large"),
  tags$br(),

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
