% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Select}
\alias{Select}
\title{Select}
\usage{
Select(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Skeleton/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

# Load Settings
options <- c("Option A", "Option B", "Option C", "Option D", "Option E")
values <- c("1", "2", "3", "4", "5")

ui <- fluidPage(

  # Application title
  titlePanel("Select Example"),

  # Select Input ------------------------------------------------------------
  tags$text("Select Input"),
  p(),
  FormControl(
    fullWidth = T,
    InputLabel(id = "demo-simple-select-label", "Election"),
    Select.shinyInput("text1",
                      #labelId = "demo-simple-select-label",
                      spacing = 2,
                      type = "required",
                      # value = 10, # Default value set to 10
                      label = "Election",
                      fullWidth = T,
                      MenuItem(value = values[1], options[1]),
                      MenuItem(value = values[2], options[2]),
                      MenuItem(value = values[3], options[3]),
                      #helperText="Some important text",
                      multiline = F
    )
  ),
  textOutput("textValue1"),
  tags$br(),

)

server <- function(input, output, session) {

  # Output: Select Input
  output$textValue1 <- renderText({
    sprintf("Value: \%s", input$text1)
  })

}

# Run the application
shinyApp(ui = ui, server = server)
}
