% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{SignInPage}
\alias{SignInPage}
\title{SignInPage}
\usage{
SignInPage(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-sign-in-page/ for more information.
}
