% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Switch}
\alias{Switch}
\title{Switch}
\usage{
Switch(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Switch/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("Switch Example"),

  # Switch Input------------------------------------------------------------------
  tags$text("Switch Input"),
  p(),
  FormControl(FormLabel("Connectors Set"),
              FormGroup(
                Switch.shinyInput("text4", value = T, name = "Connector", color="success"),
                Switch.shinyInput("text4", value = T, label = "Connector", color="warning"),
                Switch.shinyInput("text4", value = T, label = "Connector", color="secondary"))
  ),
  textOutput("textValue4"),
  tags$br(),

)

server <- function(input, output, session) {

  #Output: Switch Input
  output$textValue4 <- renderText({
    sprintf("Value: \%s", input$text4)
  })

}

shinyApp(ui, server)

}
