% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputs.R
\name{TextField.shinyInput}
\alias{TextField.shinyInput}
\title{TextField}
\usage{
TextField.shinyInput(inputId, ..., value = defaultValue)
}
\arguments{
\item{inputId}{the shiny input id}

\item{...}{args to pass to element}

\item{value}{the default value of the element}
}
\description{
Visit the documentation at https://mui.com/api/Text-Field/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("TextField Example"),

  # TextField Input -----------------------------------------------------------
  tags$text("TextField Input"),
  p(),
  TextField.shinyInput("text3",
                       fullWidth = F,
                       label = "Trial",
                       value = "", #Default
                       variant = "outlined",
                       helperText = "Some important text",
                       multiline = F),
  textOutput("textValue3"),
  tags$br(),

)

server <- function(input, output, session) {

  #Output: TextField Input
  output$textValue3 <- renderText({
    sprintf("Value: \%s", input$text3)
  })

}

shinyApp(ui, server)
}
