% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{ThemeProvider}
\alias{ThemeProvider}
\title{ThemeProvider}
\usage{
ThemeProvider(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Theme-Provider/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("ThemeProvider Example"),

  # ThemeProvider -----------------------------------------------------------
  tags$text("ThemeProvider"),
  ThemeProvider(
    theme = list(palette = list(primary = list(main = "#01cfca", contrastText = "#FFFFFF"))),
    Typography(variant = "h1", "H1", color = "primary"),
    tagList(
      Typography(variant="h1", "H1"),
      Typography(variant="h2", "H2"),
      Typography(variant="body1", "Body 1"),
    )
  ),
  tags$br(),

)

server <- function(input, output, session) {

}

shinyApp(ui, server)
}
