% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{Typography}
\alias{Typography}
\title{Typography}
\usage{
Typography(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/Typography/ for more information.
}
