% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useActivePage}
\alias{useActivePage}
\title{useActivePage}
\usage{
useActivePage(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-page-container/ for more information.
}
