% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useAutocomplete}
\alias{useAutocomplete}
\title{useAutocomplete}
\usage{
useAutocomplete(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/base-ui/react-autocomplete/hooks-api/#use-autocomplete/ for more information.
}
