% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useDialogs}
\alias{useDialogs}
\title{useDialogs}
\usage{
useDialogs(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-use-dialogs/ for more information.
}
