% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useLocalStorageState}
\alias{useLocalStorageState}
\title{useLocalStorageState}
\usage{
useLocalStorageState(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-persistent-state/ for more information.
}
