% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useMediaQuery}
\alias{useMediaQuery}
\title{useMediaQuery}
\usage{
useMediaQuery(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/material-ui/react-use-media-query/ for more information.
}
