% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useNotifications}
\alias{useNotifications}
\title{useNotifications}
\usage{
useNotifications(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/toolpad/core/react-use-notifications/ for more information.
}
