% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{usePagination}
\alias{usePagination}
\title{usePagination}
\usage{
usePagination(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/material-ui/react-pagination/#usepagination for more information.
}
