% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TextField}
\alias{TextField}
\title{TextField}
\usage{
TextField(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/api/TextField/ for more information.
}
