% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{TextField}
\alias{TextField}
\title{TextField (NO ESTÁ)}
\usage{
TextField(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/material-ui/api/text-field/ for more information.
}
\examples{
library(shiny)
library(custom4mui)

ui <- fluidPage(

  # Application title
  titlePanel("TextField Example"),

  # TextField Input -----------------------------------------------------------
  tags$text("TextField Input"),
  p(),
  TextField.shinyInput("text3",
                       fullWidth = F,
                       label = "Trial",
                       value = "", #Default
                       variant = "outlined",
                       helperText = "Some important text",
                       multiline = F),
  textOutput("textValue3"),
  tags$br(),

)

server <- function(input, output, session) {

  #Output: TextField Input
  output$textValue3 <- renderText({
    sprintf("Value: \%s", input$text3)
  })

}

shinyApp(ui, server)
}
