% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{useScrollTrigger}
\alias{useScrollTrigger}
\title{useScrollTrigger}
\usage{
useScrollTrigger(...)
}
\arguments{
\item{...}{args to pass to element}
}
\description{
Visit the documentation at https://mui.com/material-ui/react-app-bar/#usescrolltrigger-options-trigger/ for more information.
}
